+++ cluster_ilb_routing/providers.tf
@@ -24,7 +24,7 @@ terraform {
     }
     google = {
       source  = "hashicorp/google"
-      version = ">= 5.0"
+      version = ">= 5.0, < 8.0"
     }
   }
 }